Minor code updates
